requirements.txt
Flask
Flask-SSLify
Flask-BasicAuth
markdown
psycopg2-binary
PyGithub
gitpython